// src/service/file/schemas/smartFolderTreeResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { SmartFolderTreeResultSmartFolderApiType } from './smartFolderTreeResultSmartFolderApiType';
import type { SmartFolderTreeResultRootType } from './smartFolderTreeResultRootType';

/**
 * 경로 표시용 리스트
 * @nullable
 */
export type SmartFolderTreeResult = {
  /** 스마트폴더 자료의 id */
  id: number;
  /** api 요청하는 구분(Photo:사진스마트폴더, EducationalData:자료스마트폴더, UserFolder:내폴더) */
  smartFolderApiType: SmartFolderTreeResultSmartFolderApiType;
  /** 스마트폴더의 루트 종류(주로 작업) */
  rootType: SmartFolderTreeResultRootType;
  /** 상위 스마트폴더 자료의 id */
  parentSmartFolderItemId: number;
  /** 자료의 물리적 키. (DriveItem.key */
  driveItemKey: string;
  /**
   * 자료의 폴더 뎁스.
   * @minimum 1
   * @maximum 10
   */
  depth: number;
  /**
   * 폴더 이름
   * @maxLength 255
   */
  name: string;
  /** 사용자가 이름을 바꾸거나 변경할 수 있는지 여부. */
  userEditable: boolean;
  /**
   * 하위의 폴더 리스트
   * @nullable
   */
  subFolders?: SmartFolderTreeResult[] | null;
} | null;