// src/service/core/schemas/taskExampleResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Core API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

export interface TaskExampleResult {
  /** 썸네일 표시에 들어갈 제목 */
  title: string;
  /** 썸네일 표시에 들어갈 글 */
  contents: string;
  /** 썸네일용 url */
  thumbImageUrl: string;
  /** 전체 내용 url */
  fullImageUrl: string;
}